pyproject.toml
[project]
name = "obsidian-gate"
version = "0.1.0"
description = "Publish Obsidian vaults to the web with filtered content for players."
authors = [{ name="Francisco Rodrigues", email="francisco.rodrigues0908@gmail.com" }]
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "pyyaml",
    "mistune",
    "mkdocs",
    "jinja2"
]

[project.scripts]
og = "obsidian_gate.cli:main"
